Rename cloud_bdii package to cloud_info to avoid clashes.
