+++ cloud_info/exceptions.py
