+++ cloud_info/providers/opennebula.py
@@ -9,9 +9,9 @@ import string
 import urllib2
 import xml.etree.ElementTree as xee
 
-from cloud_bdii import exceptions
-from cloud_bdii import providers
-from cloud_bdii import utils
+from cloud_info import exceptions
+from cloud_info import providers
+from cloud_info import utils
 
 
 class OpenNebulaBaseProvider(providers.BaseProvider):
