+++ cloud_info/providers/openstack.py
@@ -1,6 +1,6 @@
-from cloud_bdii import exceptions
-from cloud_bdii import providers
-from cloud_bdii import utils
+from cloud_info import exceptions
+from cloud_info import providers
+from cloud_info import utils
 
 
 class OpenStackProvider(providers.BaseProvider):
