+++ cloud_info/tests/data.py
