+++ cloud_info/tests/one.imagepool.info.xml
