+++ cloud_info/tests/one.templatepool.info.xml
