+++ cloud_info/tests/utils.py
