+++ cloud_info/utils.py
