+++ rpm/cloud-info-provider-indigo.spec
@@ -42,7 +42,7 @@ rm -rf $RPM_BUILD_ROOT
 
 %files
 %defattr(-,root,root,-)
-%{python_sitelib}/cloud_bdii*
+%{python_sitelib}/cloud_info*
 %{python_sitelib}/cloud_provider*
 /usr/bin/cloud-info-provider-indigo-service
 %config /etc/cloud-info-provider-indigo/
