+++ setup.cfg
@@ -23,11 +23,11 @@ setup-hooks =
 
 [files]
 packages =
-    cloud_bdii
+    cloud_info
 data_files =
     /etc/cloud-info-provider-indigo = etc/*
 
 [entry_points]
 
 console_scripts =
-    cloud-info-provider-indigo-service = cloud_bdii.core:main
+    cloud-info-provider-indigo-service = cloud_info.core:main
